adk-blogger/requirements.txt
google-adk
python-dotenv
mcp
pytrends
